F2CTrans: Invert dims indices for column-major; no index shift
The index offset for loop variables is implicitly translated in
the code generator for loop ranges, meaning we can get away with
shifting the expected indices ni the test. This isn' particularly
rigorous, since it means any use of explicit loop indices in
expressions will cause pain, but it should work for now.