Symbolics: Enable :class:`Array` with no dimensions, but shape
In subroutine calls we often omit the index ranges of arrays, making
the variable look like a scalar. Using the declaration info we can
figure out that they are not (we know the shape), but we need to force
sympy to accept them as functions with empty `.args` and override the
pretty-printing/codegen.