Symbolics: Ensure members share a cache and do dim replacement on-the-fly
Plus a few other fixes for the EC_PHYS+SCA mode. The key thing to note
here is that, since member routines share a declaration section they
also need to share meta-data (eg. types and shapes) and thus a cache.

In turn this implies the shape-sensitive dim replacement in SCA needs
to be done in one go per shared-cache scope, as the replacement adjusts
the shape, which it depends on.