CLAWTransformation: Switch to model-data based CLAW SCA API
Implements a new approach to using the CLAW API where we mark all
demoted variabels with `!$claw model-data` instead of listing
"claw scalar" variables in the sca clause. This relies on a config
file that describes our NPROMA data layout and allows us to use separate
variables for dimension sizes and iterations spaces.