OMNI: Add a specific OMNI hack, where we post-hoc insert includes
Since OMNI fails, if it can't resolve includes, we add a special
backdoor to add `--omni_includes` to the `loki-transform.py` script.
We thus rely on the fact that missing incldudes in the first preprocessing
pass cannot find the header files and thus let the directive through. On
the second CPP pass initiated by OMNI, however, we now fail unless we
tell OMNI where the headers are.

This is clearly a hack and should really be fixed by allowing `#include`
statements through the initial CPP layer, but I haven't figured out
which magic combination of "pretty please" to give to `pcpp` yet to make
it so.