SCC: Deal with loops constrained by calls when vector loop hoisting
Outer loops that contain vector-level recursive subroutine calls
create a big problem that we now deal with in a slightly more elegant
way. This is still a bit fudged, since we don't have full dependency
analysis, but it should hopefully get us a bit further. Also enabled
proper testing of this.