CLOUDSC: Temporarily throttle CLOUDSC test to single-thread builds
This is due to a nasty race condition for the OMNI and OFP frontends
on the intermediate file for CPP preprocessing. To ensure test builds
are not taking forever, we switch to HDF% usage over Serialbox.

A proper solution for this will be to point the LOKI_TMP_DIR to
somewhere in the respective build directory for each loki invocation,
but this will come with the next round of clean ups. Until then this
works.