SCC: Ensure safe array demotion for local arrays
This now explicitly checks that we do not demote arrays that buffer
values between vector sections. For this we split the demotion into
an analysis step and an application, so that we may not interfere
with the vector section wrapping, but can use the decomposed "section"
blocks for the analysis step.