Remove __init__ for DataType base class due to behaviour change in Python 3.11
Enum values do now have the type of mixin classes, which DataType is for BasicType. Consequently, instantiation fails and
the class does not work in Python 3.11 and later. Some of that behaviour is also buggy, see the discussion in
https://discuss.python.org/t/inconsistent-behavior-with-python-3-11-enum-mixin-class-behavior/24613/4

Removing the init routine restores the desired behaviour, although we do now no longer ensure that the base class has not
been instantiated directly.