Graphical Representation of IR using graphviz
The IR is "viewable" via pprint, unfortunally this discards the inner
graph structure. For exploration and interaction a graph view is
reasonable. A Visualizer class is given derived from the Visitor class,
now logical statements are shown as nodes, with corresponding label
(source code). The label is automatically wrapped to enhance easy
visability.

Main implementation detail is the private  "__add_node" member function.
The graph is walked through and current node is added, their
respective parent and formating is done, via a simplified copy of
"format_node".