Testing for pretty_visualize
For four simplistic fortran programs indepth testing of correct graph,
node and edge creation is done. The visualizer test tries to disregard
potetntial formatting changes and focus on correctnes of the graph. A
simplistic test if graphs can be written to the file system is done as
well. Last but not least a text representation of the graph is parsed
via regex and then compared to the nodes/edges list from the Visualizer.