Add unique ID to Node: _uid
See Issue #123, copy follows:

Currently, the only mechanism for differentiating two string-identical
but distinct nodes is the .source property. However this property is only
instantiated for parsed nodes; if two identical but distinct nodes are
generated in a transformation there is no method (apart from the object
id) to tell them apart.

The above could be remedied by adding a uid attribute to the Node class.
The uid property would be generated every time a node is created. This
means that even copies of the same node would have different values of
uid, unless overridden via a constructor argument.