Transformations: Rewrite Associate resolver as an in-place mapper
Instead of finding all symbols and substituting them with their
inverse-association, we now simply run over each symbol, find its
inverse from the `.scope` and apply in-place. This requires the
substitution to be written as a symbol mapper.

As an addition feature, we can now do partial resolution in associate
bodies, where only a select sub-region of the code has its associated
symbols resolved. I've added a test to this extend.