Inline: Fix rescoping of intrinsic procedure symbols in elementals
This caused problems by a race condition, where the elemental scope
that was the original scope of intrisincs could be re-build, making
the scope weakref invalid before the final catch-all rescoping.

To fix this, I'm explicitly foxing the function body to be rescoped,
before it gets inserted. This was problematic, as intrisic procedure
symbols were not updated correctly, so I enforce indiscriminant
rescoping for intrisic procedure symbols to the given, closest scope
in `AttachScopesMapper`.