Analyse: Add DataflowAnalysis and AbstractDFA
This commit adds AbstractDFA and the concrete implementation of
`DataflowAnalysis` (actually a live-variable analysis).
DataflowAnalysis is simply a rename from the user's
perspective. Internally, everything has been brought under one
class (with the use of inner classes) and makes use of AbstractDFA.

The goal is that other forms of DFA will follow and use this pattern as a common interface.