Sanitise: Fix rescoping of symbols for nested associates
When nested associates are about to be removed, we need to account
for that when updating the symbol-scope (a priori). This new logic
does this by counting out the nested associates and picking the
appropriate one, according to the septh of the symbol in the nest.

I've extended the stat-func test, where this most often triggers.