JIT: Workaround for f90wrap issue in build test
For some reason, the parameter `jprb` local to the module is not
imported from the module in the outer interfaces that f90wrap
generates, leaving a naked symbol in one of the interface procedures.

This workaround resolves this by using vanilla constants instead, but
we should eventually extend the tests to cover this again and
potentially report/fix this upstream.