Transformation: Add option for call-replacement for `!$loki remove`
This now allows a replacement subroutine call with a custom message
to be added afetr the replacement comment for pragma-based code
removal via `!$loki remove`. As an option, we can also insert a custom
import, which will only be done if replacement was performed and if
the import does not yet exist.